pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "realitydefender>=0.1.9",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "streamlit>=1.47.1",
]
